requirements.txt
faunadb==4.0.1
Flask==1.1.2
Flask_Bootstrap4==4.0.2
pytz==2020.4
flask_bootstrap==3.3.7.1
python-decouple==3.4
